refactor: findMatchCount 내 긴 if문 제거 및 LottoRank 리팩토링

+++ src/main/java/lotto/domain/Lotto.java
@@ -39,8 +39,8 @@ public class Lotto {
         }
     }
 
-    public long countMatchNumber(Lotto winningNumbers) {
-        return lottoNumbers.stream()
+    public int countMatchNumber(Lotto winningNumbers) {
+        return (int) lottoNumbers.stream()
                 .filter(winningNumbers::contains)
                 .count();
     }
